ils font les frais partie 2

+++ meuporg-server/source/Accounts.hpp
@@ -27,8 +27,8 @@ struct Client
     std::string username = "";
 
     PlayerInput inputs = {false, false, false, false};
-    unsigned int lastPacketIdReceived = 0;
-    unsigned int lastPacketIdSent = 0;
+    unsigned long long lastPacketIdReceived = 0;
+    unsigned long long lastPacketIdSent = 0;
     unsigned int lostPackets = 0;
 };
